requirements.txt
python-telegram-bot==13.0
django==3.1.2
django-environ==0.4.5